requirements.txt
flask==3.0.3
python-dotenv==1.0.1
google-generativeai==0.7.2
langchain==0.3.3
langchain-google-genai==2.0.4
